+++ libclamav/cpio.c
@@ -124,7 +124,7 @@ int cli_scancpio_old(cli_ctx *ctx)
 	if(hdr_old.namesize) {
 	    hdr_namesize = EC16(hdr_old.namesize, conv);
 	    namesize = MIN(sizeof(name), hdr_namesize);
-	    if (fmap_readn(*ctx->fmap, &name, pos, namesize) != namesize) {
+	    if ((uint32_t)fmap_readn(*ctx->fmap, &name, pos, namesize) != namesize) {
 		cli_dbgmsg("cli_scancpio_old: Can't read file name\n");
 		return CL_EFORMAT;
 	    }
@@ -202,7 +202,7 @@ int cli_scancpio_odc(cli_ctx *ctx)
 	}
 	if(hdr_namesize) {
 	    namesize = MIN(sizeof(name), hdr_namesize);
-	    if (fmap_readn(*ctx->fmap, &name, pos, namesize) != namesize) {
+	    if ((uint32_t)fmap_readn(*ctx->fmap, &name, pos, namesize) != namesize) {
 		cli_dbgmsg("cli_scancpio_odc: Can't read file name\n");
 		return CL_EFORMAT;
 	    }
@@ -276,7 +276,7 @@ int cli_scancpio_newc(cli_ctx *ctx, int crc)
 	}
 	if(hdr_namesize) {
 	    namesize = MIN(sizeof(name), hdr_namesize);
-	    if (fmap_readn(*ctx->fmap, &name, pos, namesize) != namesize) {
+	    if ((uint32_t)fmap_readn(*ctx->fmap, &name, pos, namesize) != namesize) {
 		cli_dbgmsg("cli_scancpio_newc: Can't read file name\n");
 		return CL_EFORMAT;
 	    }
