+++ libclamav/prtn_intxn.c
@@ -51,13 +51,13 @@ int prtn_intxn_list_check(prtn_intxn_list_t* list, unsigned *pitxn, off_t start,
         (*pitxn)--;
 
         if (start > check_node->Start) {
-            if (check_node->Start+check_node->Size > start) {
+            if (check_node->Start+check_node->Size > (unsigned long)start) {
                 ret = CL_VIRUS;
                 break;
             }
         }
         else if (start < check_node->Start) {
-            if (start+size > check_node->Start) {
+            if (start+size > (unsigned long)(check_node->Start)) {
                 ret = CL_VIRUS;
                 break;
             }
